Move api to own component to minimize api calls, fix image alt warnings
