+++ src/components/Person/Item/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
-let PersonItem = ({ person: { jobTitle, headshot } }) =>
+let PersonItem = ({ person: { jobTitle, headshot, firstName, lastName } }) =>
   <div className="person" >
-    <img className="person__img" src={headshot.url} />
+    <img className="person__img" src={headshot.url} alt={`${firstName} ${lastName}`} />
     <h2 className="person__content" >{jobTitle}</h2>
   </div>
